+++ extensions/timers/src/backend/timerBackend.ts
@@ -44,9 +44,7 @@ async function startTimer({
     environment.assetsPath + "/" + (selectedSound === "default" ? prefs.selectedSound : selectedSound)
   }`;
   const cmdParts = [`sleep ${timeInSeconds}`];
-  cmdParts.push(
-    `if [ -f "${masterName}" ]; then osascript -e 'display notification "Timer \\"${timerName}\\" complete" with title "Ding!"'`,
-  );
+  cmdParts.push(`if [ -f "${masterName}" ]; then open -b ca.nathanyeung.TimersNotifHelper --args "${timerName}"`);
   const afplayString = `afplay "${selectedSoundPath}" --volume ${prefs.volumeSetting.replace(",", ".")}`;
   if (prefs.selectedSound === "speak_timer_name") {
     cmdParts.push(`say "${timerName}"`);